requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Machine Learning / Embeddings
torch==2.3.1
torchvision==0.18.1
transformers==4.44.2

# Google Cloud 
google-cloud-sqlcommenter==2.0.0

# Utils
python-multipart==0.0.9
pydantic==2.9.2


